\name{baseline}
\alias{baseline}

\title{Stratification effects on baseline functions}
\description{
This function allows one to add a simple baseline stratification term to the generalized AFT model.
}
   
\usage{
baseline(...)
}

\arguments{
   \item{...}{stratification variables to be entered; see the example in \code{\link{frailtyGAFT}}.}
}

\references{
Zhou, H., Hanson, T., and Zhang, J. (2020). spBayesSurv: Fitting Bayesian Spatial Survival Models Using R. \emph{Journal of Statistical Software}, 92(9): 1-33.

Zhou, H., Hanson, T., and Zhang, J. (2017). Generalized accelerated failure time spatial frailty model for arbitrarily censored data. \emph{Lifetime Data Analysis}, 23(3): 495-515..
}


\author{
Haiming Zhou and Timothy Hanson
}


