\name{spBayesSurv-package}
\alias{spBayesSurv-package}
\alias{spBayesSurv}
\docType{package}
\title{
Spatial Copula Modelling to Bayesian Survival Analysis
}
\description{
This package provides Bayesian model fitting for several survival models 
including spatial copula linear dependent Dirichlet process mixture model, 
anova Dirichlet process mixture model, proportional hazards model and marginal spatial proportional hazards model.
}
\details{
\tabular{ll}{
Package: \tab spBayesSurv\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-06-10\cr
License: \tab GPL (>= 2)\cr
}
This package provides functions for fitting several Bayesian survival models including 
             \code{spCopulaDDP} for spatial copula linear dependent Dirichlet process mixture model, \code{anovaDDP} 
             for anova Dirichlet process mixture, \code{indeptCoxph} for proportional hazards model, and 
             \code{spCopulaCox} for marginal spatial proportional hazards model.
}
\author{
Haiming Zhou <zhouh@email.sc.edu> and Tim Hanson <hansont@stat.sc.edu>
}
\references{
De Iorio, M., Johnson, W. O., Mueller, P., and Rosner, G. L. (2009). Bayesian nonparametric nonproportional hazards survival modeling. 
Biometrics, 65(3), 762-771. 

Zhou, H., Hanson, T. and Knapp, R. (2014+). A Spatial Copula Approach to Fully Bayesian Nonparametric Survival Analysis. 
In preparation for publication.
}
\keyword{ package }
